@@ -16,38 +16,38 @@
   },
   "dependencies": {
     "compression": "^1.6.2",
-    "cross-env": "^3.2.4",
+    "cross-env": "^4.0.0",
     "es6-promise": "^4.1.0",
     "express": "^4.15.2",
+    "extract-text-webpack-plugin": "^2.1.0",
     "firebase": "^3.7.2",
     "lru-cache": "^4.0.2",
     "serve-favicon": "^2.4.1",
-    "vue": "^2.2.4",
-    "vue-router": "^2.3.0",
-    "vue-server-renderer": "^2.2.4",
-    "vue-style-loader": "^2.0.4",
-    "vuex": "^2.2.1",
-    "vuex-router-sync": "^4.1.2",
-    "webpack-merge": "^4.0.0"
+    "vue": "^2.3.0-beta.1",
+    "vue-router": "^2.5.0",
+    "vue-server-renderer": "^2.3.0-beta.1",
+    "vuex": "^2.3.1",
+    "vuex-router-sync": "^4.1.2"
   },
   "devDependencies": {
     "autoprefixer": "^6.7.7",
     "buble": "^0.15.2",
     "buble-loader": "^0.4.1",
-    "css-loader": "^0.27.3",
-    "file-loader": "^0.10.1",
+    "css-loader": "^0.28.0",
+    "file-loader": "^0.11.1",
     "friendly-errors-webpack-plugin": "^1.6.1",
-    "html-webpack-plugin": "^2.28.0",
     "rimraf": "^2.6.1",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.1",
     "sw-precache-webpack-plugin": "^0.9.1",
     "url-loader": "^0.5.8",
-    "vue-loader": "^11.1.4",
-    "vue-ssr-webpack-plugin": "^1.0.2",
-    "vue-template-compiler": "^2.2.4",
+    "vue-loader": "^12.0.0",
+    "vue-style-loader": "^3.0.0",
+    "vue-template-compiler": "^2.3.0-beta.1",
     "webpack": "^2.2.1",
+    "webpack-merge": "^4.0.0",
     "webpack-dev-middleware": "^1.10.1",
-    "webpack-hot-middleware": "^2.17.1"
+    "webpack-hot-middleware": "^2.17.1",
+    "webpack-node-externals": "^1.5.4"
   }
 }
