@@ -3,34 +3,24 @@ import Router from 'vue-router'
 
 Vue.use(Router)
 
-// We are using Webpack code splitting here so that each route's associated
-// component code is loaded on-demand only when the route is visited.
-// It's actually not really necessary for a small project of this size but
-// the goal is to demonstrate how to do it.
-//
-// Note that the dynamic import syntax should actually be just `import()`
-// but buble/acorn doesn't support parsing that syntax until it's stage 4
-// so we use the old System.import here instead.
-//
-// If using Babel, `import()` can be supported via
-// babel-plugin-syntax-dynamic-import.
-
-const createListView = name => () =>
-  System.import('../views/CreateListView').then(m => m.createListView(name))
+// route-level code splitting
+const createListView = id => () => System.import('../views/CreateListView').then(m => m.default(id))
 const ItemView = () => System.import('../views/ItemView.vue')
 const UserView = () => System.import('../views/UserView.vue')
 
-export default new Router({
-  mode: 'history',
-  scrollBehavior: () => ({ y: 0 }),
-  routes: [
-    { path: '/top/:page(\\d+)?', component: createListView('top') },
-    { path: '/new/:page(\\d+)?', component: createListView('new') },
-    { path: '/show/:page(\\d+)?', component: createListView('show') },
-    { path: '/ask/:page(\\d+)?', component: createListView('ask') },
-    { path: '/job/:page(\\d+)?', component: createListView('job') },
-    { path: '/item/:id(\\d+)', component: ItemView },
-    { path: '/user/:id', component: UserView },
-    { path: '/', redirect: '/top' }
-  ]
-})
+export function createRouter () {
+  return new Router({
+    mode: 'history',
+    scrollBehavior: () => ({ y: 0 }),
+    routes: [
+      { path: '/top/:page(\\d+)?', component: createListView('top') },
+      { path: '/new/:page(\\d+)?', component: createListView('new') },
+      { path: '/show/:page(\\d+)?', component: createListView('show') },
+      { path: '/ask/:page(\\d+)?', component: createListView('ask') },
+      { path: '/job/:page(\\d+)?', component: createListView('job') },
+      { path: '/item/:id(\\d+)', component: ItemView },
+      { path: '/user/:id', component: UserView },
+      { path: '/', redirect: '/top' }
+    ]
+  })
+}
