@@ -1,6 +1,5 @@
 <template>
   <div class="news-view">
-    <spinner :show="loading"></spinner>
     <div class="news-list-nav">
       <router-link v-if="page > 1" :to="'/' + type + '/' + (page - 1)">&lt; prev</router-link>
       <a v-else class="disabled">&lt; prev</a>
@@ -20,17 +19,13 @@
 </template>
 
 <script>
-import Spinner from './Spinner.vue'
-import Item from './Item.vue'
-import { watchList } from '../store/api'
-
-let isInitialRender = true
+import { watchList } from '../api'
+import Item from '../components/Item.vue'
 
 export default {
   name: 'item-list',
 
   components: {
-    Spinner,
     Item
   },
 
@@ -39,18 +34,11 @@ export default {
   },
 
   data () {
-    const data = {
-      loading: false,
-      transition: 'slide-up',
-      // if this is the initial render, directly render with the store state
-      // otherwise this is a page switch, start with blank and wait for data load.
-      // we need these local state so that we can precisely control the timing
-      // of the transitions.
-      displayedPage: isInitialRender ? Number(this.$store.state.route.params.page) || 1 : -1,
-      displayedItems: isInitialRender ? this.$store.getters.activeItems : []
+    return {
+      transition: 'slide-right',
+      displayedPage: Number(this.$store.state.route.params.page) || 1,
+      displayedItems: this.$store.getters.activeItems
     }
-    isInitialRender = false
-    return data
   },
 
   computed: {
@@ -91,7 +79,7 @@ export default {
 
   methods: {
     loadItems (to = this.page, from = -1) {
-      this.loading = true
+      this.$bar.start()
       this.$store.dispatch('FETCH_LIST_DATA', {
         type: this.type
       }).then(() => {
@@ -104,7 +92,7 @@ export default {
           : to > from ? 'slide-left' : 'slide-right'
         this.displayedPage = to
         this.displayedItems = this.$store.getters.activeItems
-        this.loading = false
+        this.$bar.finish()
       })
     }
   }
@@ -143,11 +131,11 @@ export default {
     padding 0
     margin 0
 
-.slide-left-enter, .slide-right-leave-active
+.slide-left-enter, .slide-right-leave-to
   opacity 0
   transform translate(30px, 0)
 
-.slide-left-leave-active, .slide-right-enter
+.slide-left-leave-to, .slide-right-enter
   opacity 0
   transform translate(-30px, 0)
 
